# spec/openapi.yaml
openapi: 3.1.0
info:
  title: OpenCheckout API
  version: "2026-04-24"
  summary: Headless checkout primitives for merchants, widgets, and enterprise backends.
  description: |
    OpenCheckout exposes a minimal set of resource-oriented checkout primitives
    (checkoutSessions, payments, addresses) modeled after Google AIP conventions
    (AIP-121, AIP-132, AIP-134, AIP-158, AIP-193).

    **Versioning** (ADR-011): Date-stamped via the `OpenCheckout-Version` header.
    Support window ≥ 12 months; responses echo the resolved version + default +
    optional `Sunset` header (RFC 8594).

    **Errors** (ADR-010): RFC 7807 Problem Details at
    `application/problem+json` with stable `code`, `category`, `retryable`,
    localized `userMessage` (ko/en/ja), `requestId`, optional `cause` chain.

    **Authentication** (ADR-004): Three layers — L1 API Key
    (`oc_{env}_{scope}_{body}`, Bearer), L2 Session JWT (ES256 + DPoP
    RFC 9449), L3 mutual TLS. Scope model matches OAuth2 fine-grained scopes.

    **Canonical enums** (ADR-019): `PaymentStatus` (7 values), `OrderStatus`
    (9 values). Vendor vocabulary (Toss `APPROVED`, etc.) is confined to
    adapter boundaries and MUST NOT cross this API surface.
  termsOfService: https://opencheckout.dev/terms
  contact:
    name: OpenCheckout Platform
    url: https://docs.opencheckout.dev
    email: platform@opencheckout.dev
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://api.opencheckout.dev
...
security:
  - ApiKeyAuth: []
  - SessionJWT: []
    DPoP: []
  - MTLS: []
paths:
  /v1/widget/tokens:
    post:
      tags: [widget]
      summary: Issue a short-lived widget token
      operationId: createWidgetToken
      description: |
        Issues a 5-minute browser bootstrap token for the vanilla widget.
        Card data MUST NOT cross this boundary; Toss-hosted fields remain the
        only card-data surface (ADR-019 §3.12).
      x-scopes: ["checkout:create"]
      parameters:
        - $ref: '#/components/parameters/OpenCheckoutVersion'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WidgetTokenCreateRequest' }
            examples:
              merchant_page:
                value:
                  orderId: order_01HXABCDEF0123456789ABCDEF
                  origin: https://merchant.example/checkout
      responses:
        '201':
          description: Widget token issued. Never cache this response.
          headers:
            X-OC-Request-ID: { $ref: '#/components/headers/XOCRequestId' }
            Cache-Control:
              description: Always `no-store`.
              schema: { type: string, enum: [no-store] }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WidgetToken' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/public/orders/{publicId}:
    parameters:
      - $ref: '#/components/parameters/PublicOrderId'
      - $ref: '#/components/parameters/RequestId'
    get:
      tags: [widget]
      summary: Get a public order read model
      operationId: getPublicOrder
      security: []
      description: |
        Read-only, cache-friendly order projection used by browser widgets.
        The response intentionally excludes PII and tenant-private fields.
      responses:
        '200':
          description: OK.
          headers:
            X-OC-Request-ID: { $ref: '#/components/headers/XOCRequestId' }
            ETag:
              description: Weak validator for the public read model version.
              schema: { type: string }
            Cache-Control:
              description: Browser/CDN cache policy.
              schema: { type: string, example: "public, max-age=30, s-maxage=60" }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicOrder' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: "oc_{env}_{scope}_{body}"
      description: |
        Server-side API key. Prefix determines env/scope
        (`oc_live_sk_`, `oc_live_pk_`, `oc_live_rk_`, `oc_test_*`).
        Transmitted as `Authorization: Bearer <key>`.
    SessionJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        ES256 Session JWT issued to Widget/browser. TTL ≤ 5min.
        MUST be paired with `DPoP` proof header (RFC 9449).
        `none` / `HS256` rejected per RFC 8725.
    DPoP:
      type: apiKey
      in: header
      name: DPoP
      description: RFC 9449 proof-of-possession JWT for SessionJWT binding.
    MTLS:
      type: mutualTLS
      description: X.509 client certificate (enterprise endpoint only).
    OcWebhookSig:
      type: apiKey
      in: header
      name: OC-Signature
      description: |
        Outbound webhook signature header
        `OC-Signature: t=<unix>,v1=<hex>,nonce=<b64u>,kid=<key-id>`
        (see ADR-014 §3 canonical).
  parameters:
...
    PublicOrderId:
      name: publicId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
      description: Public, non-PII order projection identifier.
...
    OpenCheckoutVersion:
      name: OpenCheckout-Version
      in: header
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: "2026-04-24"
      description: API version (ISO date). Falls back to merchant account pin.
    RequestId:
      name: X-OC-Request-ID
      in: header
      required: false
      schema: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$' }
      description: Client-supplied ULID for correlation. Auto-generated if omitted.
...
  headers:
    XOCRequestId:
      description: Server-resolved request correlation ULID.
      schema: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$' }
    ResolvedVersion:
      description: API date resolved for this response.
      schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
...
    RateLimitLimit:
      description: IETF draft RateLimit header — window quota.
      schema: { type: integer }
    RateLimitRemaining:
      description: Remaining quota in current window.
      schema: { type: integer }
    RateLimitReset:
      description: Seconds until window reset.
      schema: { type: integer }
    RetryAfter:
      description: RFC 9110 retry hint (seconds).
      schema: { type: integer }
  responses:
    Problem:
      description: RFC 7807 Problem Details error.
      headers:
        X-OC-Request-ID: { $ref: '#/components/headers/XOCRequestId' }
        OpenCheckout-Version: { $ref: '#/components/headers/ResolvedVersion' }
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            invalidPostal:
              summary: Invalid Korean postal code
              value:
                type: https://docs.opencheckout.dev/errors/address/invalid-postal
                title: Invalid postal code format
                status: 422
                detail: Korean postal code must be 5 digits
                instance: /v1/addresses
                code: address.validation.postal_format
                category: validation
                retryable: false
                userMessage:
                  ko: 우편번호는 5자리 숫자여야 합니다
                  en: Postal code must be 5 digits
                  ja: 郵便番号は5桁で入力してください
                docUrl: https://docs.opencheckout.dev/errors/address/invalid-postal
                requestId: req_01HX9YZABCDEFG0123456789
    RateLimited:
      description: Rate limit exceeded.
      headers:
        X-OC-Request-ID: { $ref: '#/components/headers/XOCRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807 + OpenCheckout extension (ADR-010).
      required:
        - type
        - title
        - status
        - code
        - category
        - retryable
        - userMessage
        - docUrl
        - requestId
      properties:
        type: { type: string, format: uri, description: URI identifying the error class. }
        title: { type: string, description: Short human-readable summary. }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code:
          type: string
          pattern: '^[a-z_]+\.[a-z_]+\.[a-z_]+$'
          description: Stable dotted error code (`{domain}.{subdomain}.{reason}`).
          example: address.validation.postal_format
        category:
          type: string
          enum:
            - validation
            - authentication
            - authorization
            - rate_limit
            - payment_declined
            - external_service
            - server_error
            - resource_not_found
            - conflict
            - compliance_blocker
        retryable: { type: boolean }
        userMessage:
          type: object
          description: Localized buyer-facing messages keyed by BCP 47 tag.
          additionalProperties: { type: string }
          example: { ko: "오류가 발생했습니다", en: "An error occurred", ja: "エラーが発生しました" }
        docUrl: { type: string, format: uri }
        requestId: { type: string }
        cause:
          $ref: '#/components/schemas/Problem'
          description: Upstream / nested cause (depth ≤ 3). No PII or stack traces.
        errors:
          type: array
          description: Field-level validation issues (validation category only).
          items:
            type: object
            required: [field, code, userMessage]
            properties:
              field: { type: string }
              code: { type: string }
              userMessage:
                type: object
                additionalProperties: { type: string }
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
          format: int64
          description: Value in minor units (KRW uses whole won, 0 fractional).
          example: 89000
        currency:
          type: string
          enum: [KRW, USD, JPY]
...
    OrderStatus:
      type: string
      description: Canonical order DAG (ADR-019 §3.2).
      enum:
        - draft
        - pending_payment
        - paid
        - processing
        - label_purchased
        - in_transit
        - delivered
        - completed
        - canceled
...
    WidgetTokenCreateRequest:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: string
          minLength: 1
          description: Merchant order identifier to bind into the widget token.
        origin:
          type: string
          format: uri
          description: Merchant page origin. Normalized to scheme + host + port.
    WidgetToken:
      type: object
      required: [token, tokenType, tenantId, orderId, allowedOrigins, expiresAt]
      properties:
        token: { type: string, minLength: 16 }
        tokenType: { type: string, enum: [Bearer] }
        tenantId: { type: string }
        orderId: { type: string }
        allowedOrigins:
          type: array
          items: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
    PublicOrder:
      type: object
      required: [publicId, status, amount, updatedAt, version]
      properties:
        publicId: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        amount: { $ref: '#/components/schemas/Money' }
        updatedAt: { type: string, format: date-time }
        version:
          type: string
          description: Opaque read-model version used for weak ETags.
...